crm-scraper/requirements.txt
instaloader==4.13
requests==2.31.0
beautifulsoup4==4.12.3
supabase==2.10.0
python-dotenv==1.0.1
